refactor: Update RecordTypeBase and DataAccessorBase for null safety; modify error handling in DynamoDBService and update imports for legacy classes
