+++ finance/services/ExchangeService.ts
@@ -1,4 +1,4 @@
-import CRUDServiceBase from '@common/services/CRUDServiceBase';
+import CRUDServiceBase from '@common/services/CRUDServiceBase.old';
 import TimeUtil from '@common/utils/TimeUtil';
 
 import ExchangeDataAccessor from '@finance/services/ExchangeDataAccessor';
